updated readme v2.0.8
