+++ README.md
@@ -1,4 +1,4 @@
-# html2multipagepdf v2.0.7
+# html2multipagepdf v2.0.8
 
 Easily convert HTML templates to PDF in React.js/Next.js. Generate dynamic, server-side PDFs for invoices, reports, or documents with seamless integration into your React.js/Next.js application.
 
@@ -7,7 +7,7 @@ Easily convert HTML templates to PDF in React.js/Next.js. Generate dynamic, serv
 Using npm:
 ```sh
 $ npm i -g npm
-$ npm i @abhisek507/html2multipagepdf@2.0.7
+$ npm i @abhisek507/html2multipagepdf@2.0.8
 ```
 
 In React.js/Next.js/Vanilla JavaScript:
