+++ constants.js
@@ -1,10 +1,3 @@
-import {
-  PageQuality,
-  PageOptions,
-  OutputType,
-  PageGenerator,
-} from "./types";
-
 /**
  * @type {PageQuality}
  */
