+++ index.d.ts
@@ -1,6 +1,19 @@
-import { PageOptions, jsPDF } from "./types";
+import { jsPDF, jsPDFOptions } from "jspdf";
 
 declare module "@abhisek507/html2multipagepdf" {
+  interface PageMargin {
+    narrow: number;
+    normal: number;
+  }
+
+  interface PageOptions extends jsPDFOptions {
+    orientation: "p" | "portrait" | "l" | "landscape";
+    unit: "pt" | "px" | "in" | "mm" | "cm" | "ex" | "em" | "pc";
+    format: number[];
+    compress: boolean;
+    margin: PageMargin;
+  }
+
   export function generatePDF (
     pageSelectors: string[],
     pageOptions: PageOptions,
@@ -11,7 +24,7 @@ declare module "@abhisek507/html2multipagepdf" {
 
 export function generatePDF (
   pageSelectors: string[],
-  pageOptions: PageOptions,
+  pageOptions: jsPDFOptions,
   elementSelector: string,
   quality: number,
 ): Promise<jsPDF>;
